+++ libs/portals/my-pages/finance/src/screens/FinanceRoutes.loader.ts
@@ -1,6 +1,7 @@
 import { Query } from '@island.is/api/schema'
 import { gql } from '@apollo/client'
 import type { WrappedLoaderFn } from '@island.is/portals/core'
+import { plausibleSwapDataDomain } from '../utils/plausibleSwapDataDomain'
 
 export const GET_TAPS_QUERY = gql`
   query GetTapsQuery {
@@ -13,10 +14,20 @@ export const GET_TAPS_QUERY = gql`
   }
 `
 
+export const financeDomainLoader: WrappedLoaderFn = () => async () => {
+  try {
+    plausibleSwapDataDomain('finance')
+    return null
+  } catch {
+    return null
+  }
+}
+
 export const financeRoutesLoader: WrappedLoaderFn =
   ({ client }) =>
   async () => {
     try {
+      plausibleSwapDataDomain('finance')
       const { data } = await client.query<Query>({
         query: GET_TAPS_QUERY,
       })
